Collective5 (#94)
* game modules improvements

* importing runtime code

* adding experiments in CRI middleware

* accepting firsts round of TUs

* accepting second round of TUs

* accepting 3rd round of TUs

* addintions with annotations

* Match cam widescreen-bar and camera-offset helpers

- add_widescreen_bars 97.52% -> 100%: single early return when either
  bar object fails to load.
- fade_box_item_live_node: return 0 directly instead of assigning a dead
  local; remove_widescreen_bars 95.50% -> 98.43%.
- remove/add_camera_offsets_impl: spell the live camera lookup as a
  ternary (same shape as bgnd_nbc.c); add_camera_offsets 97.96% -> 100%,
  p_hold_camera_in_place 97.58% -> 100%, generic_victory_camera and
  p_scripted_camera improve slightly, .data string pool improves.

Stale progress comments of the now-exact functions removed. No other
function in main/cam changes.


* Match assign_obj_to_trigger by trigger definition pointer

assign_obj_to_trigger looks its trigger up by the owned definition
pointer (find_trigger_by_definition), not by trigger id: 99.88% -> 100%.
The prototype and the script dispatcher _assign_obj_to_trigger pass the
script slot as a KonquestTriggerDefinition pointer; script_functions.o is
byte-identical. The stale soft-ceiling comment is removed and the
find_trigger_by_definition "next" temporary is block-scoped.

npc_play_teleported_sound tests the attenuated volume by truthiness:
99.19% -> 99.42% (remaining delta is the pooled local-constant label).


* Fix pz_fighter_showoff_warmup2 constants and tune pz fighter code

- pz_fighter_showoff_warmup2 99.57% -> 100%: the animation weight, speed
  and blend frame were wrong; retail .sdata2 pool entries @1889/@1890/
  @1891 are 0.55f, 0.7f and 20.0f. The corrected pool order also lifts
  pz_fighter_dont_get_me and pz_fighter_whatever2 to 100% and
  pz_fighter_round_whew, pz_fighter_backflip_and_point and
  r_pz_fighter_feet3_swept_out (comment updated to 99.85%).
- pz_fighters_chomper2_fatality_in_progress 99.33% -> 99.87%: explicit
  empty case for the terminal state 4 that case 3 sets, which restores
  retail's duplicate dispatch branch.
- pz_fighters_calc_distance_to_desired_idle_pos_abs 98.09% -> 98.43%:
  retail operand order dz*dz + dx*dx.


* Match bgnd wall-hider helpers and yinyang music toggles

- bgnd_add_fx_to_hide 98.83% -> 100%: append with hidden_effects[count++].
- bgnd_remove_wall_from_hider 99.89% -> 100%: read the wall hider once
  into a local.
- yinyang_start_music 98.68% -> 100% and yinyang_stop_music 99.67% ->
  100%: test the evil-time flag in the retail branch order.
- bl_process_general_movement 99.53% -> 99.66%: keep the radian angle in
  its own local instead of reusing the arctangent heading.

Stale near-match / soft-ceiling comments of the exact functions removed.


* Match small game-code functions from the agent runs

Exact now (before -> after):
- movelist_change_style 99.05% -> 100% (nest style-slot lookup in call)
- plyr_set_vel_xz_y 99.17% -> 100% (named sine/cosine/velocity locals)
- nav_get_unit_vector_to_area 99.65% -> 100% (inline 1 - ratio blend)
- p_krypt_animate 99.87% -> 100% (volume truthiness test)
- p_save_bg_profile 99.78% -> 100% (retail branch order)
- update_x_cursor_position 97.92% -> 100% bytes (*x = *x - x_step);
  TU order was already off and is unchanged

Improved:
- drone_ai_dodge_3d_with_counter 99.85% -> 99.88% (Vec locals in
  retail stack order)
- insert_mu 98.13% -> 98.24% (device >= 2 bound as in retail compare)
- ppl_get_multi_profile_names_p1 99.04% -> 99.36% and _p2 99.04% ->
  99.55% (declaration order in the shared name-fill inline)

Progress comments updated or removed to match.


* Match runtime camera snapshot, light, file and RW-light helpers

Exact now (before -> after):
- TakeCameraSnapShot 99.87% -> 100% (block-scoped old camera frame)
- load_lights 97.71% -> 100% (walk a cursor copy of the def array)
- mk_hwfile_write_blocking 97.76% -> 100% (tell() passed directly)
- obj_set_rw_lights 98.94% -> 100% (explicit case 2 shares default)

Improved:
- insert_string_obj 98.60% -> 99.07% (separate cur_vtbl local for the
  list entry being compared); soft-ceiling note replaced by a TODO
  naming the remaining folded branch.


* Match bgnd wall-hide append and chunk-monitor explosion offset

- bgnd_add_wall_to_hide 99.13% -> 100%: append with
  walls_to_hide[hide_count++]; drop the unused index local.
- p_bgnd_launch_chunk_monitor 99.87% -> 100%: apply the +0.5f Y lift
  after the position snapshot, matching retail operand order.

No other bgnd function changes; bgnd .text 94.444% -> 94.446%.
Remove the stale near-match comments.


* Build mcardmsg with pooled read-only strings

Replace the hand-written stringBase0 array with plain literals and build
mcardmsg.o with object-scope -str reuse,pool,readonly (playbook M15), so
MWCC emits the retail @stringBase0 pool itself.

- fire_up_memcard_mesage_screen 99.18% -> 100%
- set_memcard_popup_message_{options,body,title}_text 99.47% -> 100%

.text 99.76% -> 100%; .rodata/.data/.sdata/.sbss unchanged (100%),
section sizes unchanged, main.dol SHA-1 still OK. The object stays
Matching.


* Match konquest monk validation callers

- p_transform_into_player 98.54% -> 100%: pass
  mission_state->monk_process straight into
  mission_validate_monk_process.
- full_konquest_load_from_memcard 98.17% -> 100%: resolve the monk
  through a konquest_validate_monk inline with the same shape as the
  mission helper.

No other function in either TU changes. Remove the stale near-match
comments.


* Match back_to_crouch via an animation advance-to-frame inline

- back_to_crouch 98.96% -> 100%: new static inline
  ejb_anim_advance_to_frame clamps the target to high_frame and runs
  EJB_ADVANCE_TO_FRAME.
- if_collision_slow_ani_x 97.74% -> 98.23%: declaration order fixes
  the float register swap; the r3 + mr r30 load of plyr_anim_pdata
  remains (status comment updated).

No other ejb function changes.


* Match mk_chess_blend_to_my_cell_pos code bytes

Declare the locals before initializing them (dx, dz, then original_x),
fixing the f7/f8/f9 swap: 98.49% -> 100% bytes. The function is still
not green: the whole mk_chess TU is out of retail function order,
which this change does not address. No other mk_chess function changes.


* Improve pui_update, konquest_fade_screen and p_pz_pselect

- pui_update 98.53% -> 99.29%: take pui from link->hdr in the live
  branch; size and CFG now agree, only an r4/r5 swap remains.
- konquest_fade_screen 99.36% -> 99.75%: subtract the fade-out step in
  place; size now agrees, an r3/r4 swap and string-pool relocations
  remain.
- p_pz_pselect 97.38% -> 99.82%: test the g_game_info flag before
  taking gi, and count the wait down before sleeping; size agrees and
  only named stringBase0 relocations remain.

No other function in either TU changes. Status comments updated.


* minimal changes

* adding compiler flags

---------

+++ src/game/pfxscript.c
@@ -2364,7 +2364,7 @@ void bind_to_bone(int bone_index) {
         if (active_cmdscript->mko != 0 &&
             active_cmdscript->mko->load_ctx != 0) {
             object = active_cmdscript->mko->load_ctx->bgnd_obj;
-            if ((int)object->oid == 0x1001 || (int)object->oid == 0x1002) {
+            if (object->oid == 0x1001 || object->oid == 0x1002) {
                 effect->bound_obj = object;
                 if (effect->bound_obj != 0) {
                     pfx_bind_emitter_to_obj_bone(effect, effect->bound_obj, bone_index);
